Fix double label
Label doubling was due to an issue in the animation
framework where elements are not removed from the dom
when new animation events queue before the out animation
is complete.

Tried reducing animation time and type, but no way to
out run.

Removed AnimatePresence instead, reimplemented with custom
mount check.